Fix attachments / Implement basic backpressure
* Attachments need to be base64 encoded. Fixed.

* Nothing was being done to prevent receiver events from being dropped
after the first client connection was established.  We now ensure that
the message receiver shuts down and queues any overflowed events in prep
for the next connection.  E.g. queue the messages received in the window
where we don't have a valid client connection but have already accepted
the message from the signal server.  Without this we leak events like a
sieve.

* Removed close event.  This would be undeliverable anyway.

* Increase logging

* Wrap stateful async functions with queue_async so we don't corrupt
our state machine.